COG-1197: interface Diff is removed
