+++ src/main/java/gov/ca/cwds/idm/service/diff/BooleanDiff.java
@@ -1,6 +1,6 @@
 package gov.ca.cwds.idm.service.diff;
 
-public final class BooleanDiff extends BaseDiff<Boolean> {
+public final class BooleanDiff extends Diff<Boolean> {
 
   public BooleanDiff(final Boolean oldValue, final Boolean newValue) {
     super(oldValue, newValue);
