+++ src/main/java/gov/ca/cwds/idm/service/diff/StringDiff.java
@@ -1,6 +1,6 @@
 package gov.ca.cwds.idm.service.diff;
 
-public final class StringDiff extends BaseDiff<String> {
+public final class StringDiff extends Diff<String> {
 
   public StringDiff(final String oldValue, final String newValue) {
     super(oldValue, newValue);
